Reduce code redundancy through additional template in blocking_queue (#57)
# Description
Thanks for your great work on this project, solving one of the main
painpoints for most robotics applications today with modern c++!
By skimming through it, I noticed that some code redundancy could be
removed. However this comes at the cost of having to write the type of
constructed object for the functions explicitly, as can be seen in the
examples.
What do you think?

I am not sure how to mark this PR, since [Feat] is not really applicable
in my opinion. Do you have conventions for refactorings?

## Type of change
- New feature (non-breaking change which adds functionality). Actually
its a proposal for reducing redundancy, but this was the best fit of the
options...

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests. -> tested it
with the examples and tests supplied
- [ ] If this is a new component I have added examples. -> not
applicable
- [ ] I updated the README and related documentation. -> not needed
since functionality is the same

---------

Co-authored-by: Johannes Graeter <j.graeter@dotsene.com>
Co-authored-by: Filippo Brizzi <filippobrizzi@gmail.com>