Zenoh bag player (#72)
# Description
Add bag player library and app:
* Right now the bag player is basic and all topic share the same thread.
This I already know is a problem but I wanted to have something running
end-to-end
* Add test that: create an mcap bag -> play the bag -> record the bag ->
check that the two bags are the same
* To do this I create test protofile in the bag test folder. These are
very similar to the one that already exists in the `serdes` module. I
will unify them later as I need to find a satisfying way of doing it.
* Extendend signal handler utilis to support the case of bag playback
where we want to terminate either because the app complete or via
signal, please provide better naming!