Fix some IWYU issues; fix typo in README. (#98)
# Description

* Fix some missing includes to make things compile; it looks like there
might be other instances where more diligence is needed w.r.t IWYU, but
these I stumbled upon as it made the compile hick-up.
  * The `ninja checks` in the README should really be `check`
* There is another typo in the README `cmake --preset preferred` ... the
`preferred` preset does not actually exist. Maybe `default` was meant ?
Anyway, not touching that, but you should probably change that to one of
the existing options.