[Fix: Cmake] properly set module prefix (#150)
# Description
In a previous PR
(https://github.com/olympus-robotics/hephaestus/pull/145) I fixed cmake
to add the project as prefix to the module name, to make Hephaestus
properly work when it is used as a submodule.
The fix was incomplete, as it was not accounting for all the places
where the module is used. Hopefully, now it works better.