Feat/dummy type serialization (#136)
# Description
Module **types_proto** provides mapping of **types** data structures
to proto IDL.

## Detailed description

Protobuf is used to serialize all data. Protobuf files are human
readable, and export to JSON is supported out of the box. This allows to
dump data into files and read it back later in a human readable format.

### Goals
* proto IDL files for a type are feature complete, i.e. serializing and
deserializing leads to the exact same data structure
* provide generic helper functions for serialization

### Non-goals
* similar to module **types**, **types_proto** is meant to be
generic, and project specific (non-generalizable) methods shall not be
added

### Requirements
* Types are independent of their serialization method, but the module
**types_proto** depends on module **types**.

## Type of change
- New feature (non-breaking change which adds functionality)

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.
- [ ] If this is a new component I have added examples.
- [x] I updated the README and related documentation.

---------

Co-authored-by: Filippo Brizzi <filippobrizzi@gmail.com>
Co-authored-by: Florian Tschopp <florian_tschopp@hotmail.com>