Fix/Correctly forward exceptions thrown in Spinner callback and allow for stopping (#187)
# Description
* Correctly forward exceptions thrown in `spin()`. Since `spin()` runs
in a separate thread, we need to propagate the exception message out of
`spin`.
* Exception messages are caught during `stop()`, which is usually
preceded by a call to `wait()`. `wait()` will exit if an exception is
thrown, such that it can be caught in the main thread via `stop()` right
away.
* Allow to add a return value to the callback to allow to stop the
spinner at completion, i.e. add a `StoppableCallback`
* 
## Type of change
- Bug fix
- New feature

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.
- [x] If this is a new component I have added examples.
- [ ] I updated the README and related documentation.

---------

Co-authored-by: Filippo Brizzi <filippobrizzi@gmail.com>