Add Spinner manager (#198)
# Description
The `SpinnersManager` allows to orchestrate the start and termination of
multiple spinners.

The most important functionality is the `waitAny` function which allows
to unblock as soon as one of the spinner terminates. In this way we can
proceed to stop all of the other. This is very useful in the context of
errors.