Expose proto serdes functionalities (#242)
# Description
We expose some protobuf serdes functionalities to avoid having to rely
on `ProtoAssociation`. This functionalities are in `internal` namespace
and should be used only in rare cases.
The use case is that we have one type that can match to two different
proto types. This forces as to manually specify the association in one
of the two cases.

> Note: nothing change on the public interface

Some minor changes:
* Fix a type
* Improved one error message

---------

Co-authored-by: Filippo Brizzi <brizzi@filics.eu>