[IPC] Service: add post reply callback (#255)
# Description
In the `Service`, add the option to specify a callback to be called
after the reply has been sent back.

One use case is the caller to an `ActionServer` that uses it to be
notified when the process has terminated and can destroy the temporary
Service used to receive the response back from the Server.