Fix cmake, non-exception, clang++18 build and add zstd to mcap [reopened from fork] (#267)
This PR is a re-opened version of #263 to side-step the CI issues.

# Description

As part of updating to the newest upstream changes, I enabled unit test
and example building in our deployment. This then resulted in the
changes below:
 - **fix build for the non-except compile option:** 
- Address issue where the unit test cannot pass if heph is compiled with
`DISABLE_EXCEPTIONS` since it relies on `EXPECT_THROW` in a lot of
places -> Introducing `EXPECT_THROW_OR_DEATH` in the tests
- The actual non-except `CHECK`'s didn't seem to build properly ->
changed to `ABSL_CHECK`
 - **`clang++18` build support:**
The one instance of using std::format did not compile on our system with
clang++-18. Since this is only for testing purposes anyways, I dared to
propose not compiling that part if the GLIBCXXX is not yet on version
13+. Open to discuss options, maybe there is a better way
 - **Breaking `examples` -> `format` dependency**
Building the `format` module currently requires cmake 3.23. Because we
don't need it and support it, we prefer not building the `format`
module. However recent changes added a `format`-example to the
`examples` module (which so far are exclusively IPC examples) -> so now
`examples` depends on `format`. I moved that example into the `format`
module to break that unnecessary dependency.
 - **Add zstd** for mcap compression

## Type of change

- Bug fix (non-breaking change which fixes an issue)
- Improvement