[fix] name buffer length being too small (#305)
# Description

When adding panics, a "reallocate: invalid old size" appeared in the
tests.
This was strange since no functional changes where made.
After debugging we found out, that the fixed name size in the Stacktrace
was apparantly too small.
We fix this by letting __cxa_demangle allocate memory for the words on
the heap and manage the memory ourselfs.

thanks @sithhell for the pairing on this!

## Type of change
Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.