Module - Hephaestus <-> Websocket Bridge  (#299)
![image](https://github.com/user-attachments/assets/71d572c4-8538-418a-adfb-f5cb7d6a0f37)

# Description

This module offers a websocket interface to hephaestus, enabling
foxglove studio or a similar websocket-capable client to connect to
hephaestus communication and:
  - Publish and subscribe to hephaestus topics
  - Trigger hephaestus service calls
- Monitor the hephaestus node graph (Foxglove has a built in node graph
visualizer)

At a high-level, this means for hephaestus:
- This is the equivalent of
https://github.com/foxglove/ros-foxglove-bridge, but for hephaestus
- We get "RVIZ"-like functionality and some RQT goodies (like
visualizing the IPC graph) via foxglove studio
- Creates a nice interface to connect a something like an
Android/Flutter/iOS app to the hephaestus data stream

For more details, check out the README

 ## Review Suggestions

- I know it's a lot at once. However instead of me spreading this 99%
new-code over 10 stacked PRs, such that it takes me 2 weeks to
re-assemble all the comments, I think reviewing this monster here one
header/source pair at a time, is probably more efficient for everyone.

## Module Maturity

This is a first prototype of this bridge. It does all the things
mentioned above and has a decent amount of tests, but will need some
integration work, polishing and tweaking. Feedback is welcome. See the
list at the bottom for open TODOs.

## Type of change

- New feature (non-breaking change which adds functionality)

## Limitations
- This bridge currently only supports protobuf encoding. This means that
some native foxglove studion functions, like calling service calls or
some built-in client-side messages are currently not supported because
they only work with JSON.
 
## DEPENDENCIES

- Fork of ws-protocol:
https://github.com/olympus-robotics/ws-protocol/pull/1

## Follow Up Work Planned

- [ ] Heph needs properly async service calling  /  a better RPC model
- [ ] Client side requests are processed by a worker pool. By default
this was set to 1 thread, likely for simplicity sake. However, this
meant that any client-side long-running operation blocked the entire
dataflow from the client. Going from 1 to multiple threads, fixes this
-> likely we lose client-side message ordering guarantees. Investigate
and fix if an issue
- [ ] Actually make the blacklist/whitelist settings do something,
currently they are mostly ignored.
- [ ] Test some of the advanced websocket interface functions, if they
actually work. If not, do not expose them. (Stuff like TLS and
Compression)
- [ ] Add an interface for message transfomers. The idea is that you can
register different types of Transfomers to modify the types, add fields
to the messages or throttle / combine messages, etc.

---------

Co-authored-by: Florian Tschopp <fts@voliro.com>