Conduit: Adding Input and Output facilities for nodes (#342)
- Outputs: A node can either define an explicit output
`OutputConnection<T>` or return a value from its `operator()`. If
`operator()` returns an optional, the optional will be automatically
unwrapped if it contains a value and propagated. If the optional is
empty, output propagation will be omitted. The execution will be delayed
if one of the connected inputs is full.
- Inputs: A node can define its inputs, which can be connected to
outputs. This is done by calling `node.input.connecTo(...)`. The output
can either be the implict output or an explicitly defined output. An
input has a depth and a retrieval method. The depth defines the maximal
available number of entries this input can have. The retrieval method
defines if calling `get()` on an input will block until data is
available or return immediately with an optional, signalling if there
was input available or not. The default input depth is 1 and the default
retrieval method is blocking.
- `QueuedInput<T>`: This input type is capable of queing up entries. The
retrieval is in FIFO order. It will block if no input entry is
available.
- `AccumulatedInput<T, F>`: This input type will accumulated the
propagated outputs. When the items are received, they are passed to the
accumulate function which can futher process the accumulated inputs. For
example merge them.
- Changes for `Node<Operation>`: The `trigger` function of an
`Operation` can now specify their inputs using `stdexec` algorithms like
`when_all` or `when_any`. This way, the inputs can be queried in the
concurrency framework and are forwarded to `operator()` with their
respective function signatures.

Flyby Changes:
- Adding heph::flushLogEntries, which enables deterministic testing
against log messages
 - Fixing problems with timeouts in the concurrency module


## Type of change
Please delete options that are not relevant.


- New feature (non-breaking change which adds functionality)


## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.
- [x] If this is a new component I have added examples.
- [ ] I updated the README and related documentation.