Address queue overflow in raw sub and endpoint discovery (#369)
# Description

Fixes the issue where liveliness tokens get lost due to the queue being
full and so the websocket bridge (and other discovery based tools) were
losing topics and services. This happened especially when the bridge was
the late-joiner and therefore got all liveliness tokens at once.

## Type of change
Please delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)