Feat/add leaner accumulated input interface (#388)
# Description

make interface of accumulated input leaner with an alias.
Concept for it is deliberately not added to utils/concepts.h to avoid
the dependency.

This is an API break. AFAIK this feature is not used yet so this is a
good point to change the API.

## Type of change
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.