Fix CMake compilation/linking on ARM64/Windows with CMake 4.x (#438)
# Description

This PR fixes some incorrect syntax in the `externals/CMakeLists.txt`
file (commas for args, repetition of `URL`) and ensures both absl and
protobuf are built statically and with PIC to avoid linker issues on
ARM64/Windows builds. The minimum CMake version warning is also
suppressed for libraries that do not support CMake 4.x.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.
- [x] If this is a new component I have added examples.
- [x] I updated the README and related documentation.