Feature: Adding type erased Sender (#449)
`AnySender<T>` is a type which can be used to erase the actual sender
expression type. This is useful to move sender expression into cpp
files.

This sender has three completion types:
 - `set_value(T)`: Completing with value
 - `set_error(std::exception_ptr)`: Completing if an exception occured
 - `set_stopped()`: The wrapped sender was canceled

In addition, a connected `AnySender<T>` will forward the scheduler and
stop token passed in from the environment of the receiver.

## Type of change

- New feature (non-breaking change which adds functionality)

## Checklist before requesting a review
- [x] I have performed a self-review of my code.
- [x] If it is a core feature, I have added thorough tests.
- [x] If this is a new component I have added examples.
- [x] I updated the README and related documentation.